projects/security-testing-framework/requirements.txt
PyYAML==6.0.2
requests==2.32.3
jinja2==3.1.4
pytest==8.3.5
